refactor: change date to Halloween; update readme
